+++ .env.development
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_API_URL="https://api.modumozu.com"
