+++ .env.production
@@ -1,1 +1,1 @@
-NEXT_PUBLIC_API_URL=URL
+NEXT_PUBLIC_API_URL="https://api.modumozu.com"
