+++ src/app/home/page.tsx
@@ -1,10 +1,14 @@
 "use client";
 
-import { FC } from "react";
+import { FC, useEffect } from "react";
 import UpcomingStockSection from "@/components/home/UpcomingStockSection";
 import EndedStockSection from "@/components/home/EndedStockSection";
+import { getIPOList } from "@/api/ipo";
 
 const HomePage: FC = () => {
+  useEffect(() => {
+    getIPOList();
+  }, []);
   return (
     <>
       <UpcomingStockSection />
