+++ src/constants/queryKeys.tsx
@@ -1,0 +1,5 @@
+const queryKeys = {
+  DETAIL: ["DETAIL"],
+};
+
+export default queryKeys;
